Move Coin into the Numismatics namespace
