+++ app/graphql/types/resource.rb
@@ -2,7 +2,7 @@
 module Types::Resource
   include Types::BaseInterface
   description "A resource in the system."
-  orphan_types Types::CoinType,
+  orphan_types Types::Numismatics::CoinType,
                Types::EphemeraFolderType,
                Types::FileSetType,
                Types::Numismatics::IssueType,
