+++ app/views/catalog/_members_issue.html.erb
@@ -1,6 +1,6 @@
 <table class="table table-striped datatable">
   <thead>
-    <tr><h3>Coins</h3><tr>
+    <tr><h3>Numismatics::Coins</h3><tr>
     <tr>
       <th>Thumbnail</th>
       <th>Title</th>
@@ -26,7 +26,7 @@
       <% end %>  
     <% else %>
       <tr>
-        <td>This work has no coins attached.  Click "Attach Coin" button to create new coin.</td>
+        <td>This work has no coins attached.  Click "Attach Numismatics::Coin" button to create new coin.</td>
         <td></td>
         <td></td>
         <td></td>
