+++ app/views/numismatics/coins/_file_manager_extra_tools.html.erb
