+++ app/views/numismatics/coins/_loan_fields.html.erb
