+++ app/views/numismatics/coins/_provenance_fields.html.erb
