+++ app/views/numismatics/coins/edit_fields/_append_id.html.erb
