+++ app/views/numismatics/coins/edit_fields/_find_place_id.html.erb
