+++ app/views/numismatics/coins/edit_fields/_holding_location.html.erb
