+++ app/views/numismatics/coins/edit_fields/_loan.html.erb
