+++ spec/services/manifest_builder_spec.rb
@@ -723,8 +723,8 @@ RSpec.describe ManifestBuilder do
       output = manifest_builder.build
       expect(output["@type"]).to eq "sc:Collection"
       expect(output["manifests"].length).to eq 2
-      expect(output["manifests"][0]["label"]).to eq ["Coin: 1"]
-      expect(output["manifests"][1]["label"]).to eq ["Coin: 2"]
+      expect(output["manifests"][0]["label"]).to eq ["Numismatics::Coin: 1"]
+      expect(output["manifests"][1]["label"]).to eq ["Numismatics::Coin: 2"]
     end
   end
 end
